Tests correctos

+++ src/main/java/es/upm/miw/pd/composite/Expresion.java
@@ -1,7 +1,7 @@
 package es.upm.miw.pd.composite;
 
 public abstract class Expresion {
-	
-	public abstract double operar();
+
+	public abstract int operar();
 
 }
